:recycle; Refactor for safariview and page view
#
# ==== Emojis ====
# 🐛  :bug: バグ修正
# 🤘  :metal: 機能改善
# 🐚  :shell: コマンドを実行した結果とか
# ✨  :sparkles: 部分的な機能追加
# 🎉  :tada: 盛大に祝うべき大きな機能追加
# 💉  :syringe: libraryの追加
# 📝  :memo: ドキュメント・READMEの整備
# ♻️   :recycle: リファクタリング
# 🚿  :shower: 不要な機能・使われなくなった機能の削除
# 💚  :green_heart: テストやCIの修正・改善
# 👕  :shirt: Lintエラーの修正やコードスタイルの修正
# 🚀  :rocket: パフォーマンス改善
# 🆙  :up: 依存パッケージなどのアップデート
# 🔒  :lock: 新機能の公開範囲の制限
# 👮  :cop: セキュリティ関連の改善
# 🎨  :art: デザイン修正
# 🚧  :construction: WIP 一時的なコミット

# ==== Format ====
# :emoji: Subject
#
# Commit body...

# ==== The Seven Rules ====
# 1. Separate subject from body with a blank line
# 2. Limit the subject line to 50 characters
# 3. Capitalize the subject line
# 4. Do not end the subject line with a period
# 5. Use the imperative mood in the subject line
# 6. Wrap the body at 72 characters
# 7. Use the body to explain what and why vs. how
#
# How to Write a Git Commit Message http://chris.beams.io/posts/git-commit/

# reference from http://memo.goodpatch.co/2016/07/beautiful-commits-with-emojis/
# for https://github.com/bannzai/commit_template
